+++ packages/orq-rc/src/orq_ai_sdk/models/createbudgetop.py
@@ -241,5 +241,5 @@ class CreateBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-11-03T14:12:36.898Z")
+    updated: Optional[datetime] = parse_datetime("2025-11-03T14:52:16.094Z")
     r"""The date and time the resource was last updated"""