+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetitemop.py
@@ -1320,7 +1320,7 @@ class Evaluations3(BaseModel):
 
     source: Optional[CreateDatasetItemEvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2025-11-03T14:12:48.890Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2025-11-03T14:52:27.422Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -1372,7 +1372,7 @@ class Evaluations2(BaseModel):
 
     source: Optional[EvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2025-11-03T14:12:48.889Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2025-11-03T14:52:27.421Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -1424,7 +1424,7 @@ class Evaluations1(BaseModel):
 
     source: Optional[Source] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2025-11-03T14:12:48.889Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2025-11-03T14:52:27.419Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -1498,5 +1498,5 @@ class CreateDatasetItemResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-11-03T14:12:36.462Z")
+    updated: Optional[datetime] = parse_datetime("2025-11-03T14:52:15.707Z")
     r"""The date and time the resource was last updated"""