+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetop.py
@@ -93,5 +93,5 @@ class CreateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-11-03T14:12:36.462Z")
+    updated: Optional[datetime] = parse_datetime("2025-11-03T14:52:15.707Z")
     r"""The date and time the resource was last updated"""