+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -230,7 +230,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K9510EGKGQ3E3BJVE5V5BEKP"
+        "01K953918EYV6T5B0E60ZB7TWJ"
     )
     r"""The unique identifier of the data source"""