+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K9510EGP8DHVB3HH3AFRCZ72"
+        "01K953918G5D1VGY0AZBQ5NNMV"
     )
     r"""The unique identifier of the data source"""